refactored payment system

+++ src/paymentSystem/MakePayment.java
@@ -1,5 +1,5 @@
 package paymentSystem;
 
 public interface MakePayment {
-    void pay(double amount);
+    void pay(long amount);
 }
