+++ src/paymentSystem/MakePaymentViaCreditCard.java
@@ -2,7 +2,7 @@ package paymentSystem;
 
 public class MakePaymentViaCreditCard implements MakePayment {
     @Override
-    public void pay(double amount) {
+    public void pay(long amount) {
         System.out.println(amount + " Paid via CreditCard");
     }
 }
